# pkg/web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the ISDuBA API.",
        "title": "ISDuBA API",
        "contact": {
            "name": "ISDuBA team",
            "url": "https://github.com/ISDuBA/ISDuBA",
            "email": "info@intevation.de"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/stats/critical": {
            "get": {
                "description": "Returns criticality statistics for all sources.",
                "produces": [
                    "application/json"
                ],
                "summary": "Returns criticality statistics.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Timerange start",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Timerange end",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Time step",
                        "name": "step",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/stats/critical/feed/{id}": {
            "get": {
                "description": "Returns criticality statistics for the specified feed.",
                "produces": [
                    "application/json"
                ],
                "summary": "Returns criticality statistics.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feed ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Timerange start",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Timerange end",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Time step",
                        "name": "step",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/stats/critical/source/{id}": {
            "get": {
                "description": "Returns criticality statistics for the specified source.",
                "produces": [
                    "application/json"
                ],
                "summary": "Returns criticality statistics.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Timerange start",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Timerange end",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Time step",
                        "name": "step",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/stats/cve": {
            "get": {
                "description": "Returns cve statistics for all sources.",
                "produces": [
                    "application/json"
                ],
                "summary": "Returns cve statistics.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Timerange start",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Timerange end",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Time step",
                        "name": "step",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/stats/cve/feed/{id}": {
            "get": {
                "description": "Returns cve statistics for the specified feed.",
                "produces": [
                    "application/json"
                ],
                "summary": "Returns cve statistics.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feed ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Timerange start",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Timerange end",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Time step",
                        "name": "step",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/stats/cve/source/{id}": {
            "get": {
                "description": "Returns cve statistics for the specified source.",
                "produces": [
                    "application/json"
                ],
                "summary": "Returns cve statistics.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Timerange start",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Timerange end",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Time step",
                        "name": "step",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/stats/imports": {
            "get": {
                "description": "Returns import statistics for all sources.",
                "produces": [
                    "application/json"
                ],
                "summary": "Returns import statistics.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Timerange start",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Timerange end",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Time step",
                        "name": "step",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/stats/imports/feed/{id}": {
            "get": {
                "description": "Returns import statistics for the specified feed.",
                "produces": [
                    "application/json"
                ],
                "summary": "Returns import statistics.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Feed ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Timerange start",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Timerange end",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Time step",
                        "name": "step",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/stats/imports/source/{id}": {
            "get": {
                "description": "Returns import statistics for the specified source.",
                "produces": [
                    "application/json"
                ],
                "summary": "Returns import statistics.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Timerange start",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Timerange end",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Time step",
                        "name": "step",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}